feat: Fix typo in jogador input rendering
The code changes fix a typo in the rendering of the jogador input in the Calculadora component. The conditional rendering of the space between jogador inputs was incorrect, resulting in an extra space being displayed for the first jogador. This has been fixed by changing the conditional expression to display an empty string instead of a space.